add specific GroupImport component
