+++ src/components/specific/groups/group-import/GroupImport.scss
@@ -1,0 +1,22 @@
+
+.group-import {
+  &__menu {
+    top: 35px;
+    position: absolute;
+  }
+
+  &__children-menu {
+    &--header {
+      display: flex;
+      margin: auto;
+      width: 88%;
+      border-bottom: 1px solid var(--color-silver-light);
+      padding: 6px 6px 12px 0px;
+      margin-bottom: 6px;
+    }
+
+    &--footer {
+      margin: 6px auto 0px;
+    }
+  }
+}
